requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
pybrctl==0.1.4
Pygments==2.18.0
pyroute2==0.7.12
python-pytun==2.4.1
PyYAML==6.0.2
requests==2.32.3
requests-unixsocket2==0.4.2
rich==13.9.4
rich-argparse==1.6.0
setuptools==75.6.0
urllib3==2.2.3
